move stars file within usage directory
