+++ projects/packages/forms/src/blocks/input-rating/stars.js
